+++ ft_print_hex.c
@@ -6,13 +6,13 @@
 /*   By: rpena-ro <rpena-ro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/11 16:26:00 by rpena-ro          #+#    #+#             */
-/*   Updated: 2025/12/12 13:07:31 by rpena-ro         ###   ########.fr       */
+/*   Updated: 2025/12/15 15:43:34 by rpena-ro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_print_hex(unsigned long a, const char format)
+int	ft_print_hex(unsigned long long a, const char format)
 {
 	char	*buffer;
 	int		cont;
